backend/requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
colorlog==6.7.0
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-enumfields==2.1.1
django-extensions==3.2.1
django-timezone-field==5.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
flower==1.2.0
humanize==4.4.0
idna==3.4
kombu==5.2.4
loguru==0.6.0
packaging==23.0
Pillow==9.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
pytz==2022.7
redis==4.3.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.1.0
